sobras a los formularios, sweet aletrt a los formularios y algunos estilos mas. cambio nombre de Card a CarProducto.

+++ src/components/ProductoDetalle.jsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from "react";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import "../styles/ProductoDetalle.css";
-import { dispararSweetAlertBasico } from "../assets/SweetAlert";
+import { dispararSweetAlertBasico, dispararSweetAlertTrueFalse } from "../assets/SweetAlert";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { CarritoContext } from "../contexts/CarritoContext";
 import { Button } from "react-bootstrap";
@@ -30,13 +30,18 @@ function ProductoDetalle({}) {
     })
   }, [id]);}
 
-  function dispararEliminar(){
-    eliminarProducto(id).then(()=>{
-      navigate('/productos')
-    }).catch((error)=>{
-      dispararSweetAlertBasico("Hubo un problema al agregar el producto", error, "error", "Cerrar")
-    })
-
+  async function dispararEliminar(){
+    const confirmar = await dispararSweetAlertTrueFalse('¿Estás seguro que quieres eliminar este producto?', "Esto no se puede deshacer", 'warning','Sí, Eliminar')// window.confirm('¿Estás seguro de eliminar?');
+    console.log(confirmar)
+    if (confirmar) {
+      eliminarProducto(id).then(()=>{
+        dispararSweetAlertBasico('Eliminado', 'Producto eliminado correctamente.', 'success','OK')
+        navigate('/productos')
+      }).catch((error)=>{
+        alert('Hubo un problema al eliminar el producto.');
+        dispararSweetAlertBasico("Hubo un problema al eliminar el producto", error, "error", "Cerrar")
+      })
+    }  
   }
 
 
@@ -67,9 +72,10 @@ function ProductoDetalle({}) {
         <small>{productoSeleccionado.description}</small>
         <p>{productoSeleccionado.price} $</p>
         {admin?
-          <div>
-            <Link to={"/admin/editarProducto/" + id}><Button variant="outline-warning">Editar Producto</Button></Link>
-            <Button variant="outline-danger" onClick={dispararEliminar}>Eliminar Producto</Button>
+          <div className="d-flex  flex-column ">
+            <Link to={"/admin/editarProducto/" + id}><Button className="mx-auto mb-2 " variant="outline-warning">Editar Producto</Button></Link>
+            <Button className="mx-auto mb-2 " variant="outline-danger" onClick={dispararEliminar}>Eliminar Producto</Button>
+            <Link to={"/productos"}><Button  variant="outline-success">Ir a Productos</Button></Link>
           </div> 
           :
           <div >
@@ -81,7 +87,7 @@ function ProductoDetalle({}) {
             <div className="d-flex  flex-column ">
               <Button className="mx-auto mb-2 " variant="outline-primary" onClick={funcionCarrito}>Agregar al carrito</Button>
               <Link className="mx-auto mb-2" to={"/productos/"}> <Button variant="outline-success ">Volver a Productos</Button> </Link>
-              <Link to={"/carrito"}><Button  variant="outline-warning">Ir a Carrito</Button></Link> 
+              <Link to={"/carrito"}><Button  variant="outline-warning">Ir a Carrito</Button></Link>
             </div>  
           </div>
         }
